requirements.txt
flask
python-dotenv
langchain-openai
langchain-community
langchain-core
pydantic